# [STORM-1007] Added tuple arrival rate and tuple sojourn time to QueueMetrics