I added two important metrics, namely *arrival time* and *sojourn time*, to QueueMetrics. 

Modification Summary:
1. Created a new RateTracker utility, which could report instantaneous rate upon certain event;
2. Added test cases for RateTracker;
3. Added arrival rate and sojourn time metrics to QueueMetrics; 
4. Use sampling in QueueMetrics to reduce the metric maintenance cost. 

Such metrics are very useful to performance tuning, especially for application with strong constraint on processing latency such as threat detection and smart traffic. A high sojourn time, for example, indicates that a higher degree of parallelism should be applied to the current bolt in order to reduce processing delay. 

Any comment or suggestion regarding to this pull request is welcomed. 

Thanks.