@HeartSaVioR 2015-09-03
Great work. Since it can affect performance I recommend you to include benchmark test result.
You can refer https://github.com/apache/storm/pull/521#issuecomment-106038982 to how to do it.