@HeartSaVioR 2015-09-06
@wangli1426 
Great. 
One thing which blocks me voting +1 is that I can't understand how RateTracker works. 
Could you explain how it works?