@wangli1426 2015-09-09
@HeartSaVioR 
Thanks for voting +1.

By the way, in my original design, I didn't use timer to update the histograms. Instant, the histograms were updated in ```notify()```. However, such a design requires to call ```System.currentTimeMills()``` every time when ```notify()``` is called, to determine whether an update to the histograms is needed. ```System.currentTimeMills()``` is fairly expensive to be called so frequently. So, in current implementation, I use a timer to periodically update the histograms.