@wangli1426 2015-09-17
@HeartSaVioR 

Sorry to interrupt, but may I ask when this PR could be merged? I am planing to display the new metrics on the web UI, if this PR gets merged. Thanks.