@HeartSaVioR 2015-09-17
@wangli1426 Sorry for the delay. I intended to wait for other committers to review, but since no one is reviewing I'll merge right now.